feat: find by orderCode

+++ src/main/kotlin/com/example/demo/payment/service/MoMoProcessor.kt
@@ -2,12 +2,13 @@ package com.example.demo.payment.service.external
 
 import org.springframework.stereotype.Service
 import com.example.demo.entity.PaymentProvider
+import java.math.BigDecimal
 
 @Service
 class MomoProcessor : PaymentProcessor {
     override fun getProvider(): PaymentProvider = PaymentProvider.MOMO
 
-    override fun createPayment(orderId: String, amount: Double): String {
+    override fun createPayment(orderId: String, amount: BigDecimal): String {
         // Gọi API Momo để tạo thanh toán (giả lập URL thanh toán)
         return "https://momo.com/payment?orderId=$orderId&amount=$amount"
     }
