+++ src/main/kotlin/com/example/demo/payment/service/PaymentProcessor.kt
@@ -1,8 +1,9 @@
 package com.example.demo.payment.service.external
 
 import com.example.demo.entity.PaymentProvider
+import java.math.BigDecimal
 
 interface PaymentProcessor {
     fun getProvider(): PaymentProvider
-    fun createPayment(orderId: String, amount: Double): String
+    fun createPayment(orderId: String, amount: BigDecimal): String
 }
